backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
stripe==7.10.0
razorpay==1.4.1
google-generativeai==0.3.2
supabase==2.3.4
alembic==1.13.1
python-dotenv==1.0.0
